sorting out test directory
